made code more reusable
