+++ feed/index.html
@@ -56,7 +56,7 @@
         <!--About user-->
         <div class="col-lg-4 t">
           <div class="bg-light rounded-2 p-3">
-            <h1 class="user" id="user-name">User</h1>
+            <h1 class="user" id="userName"></h1>
             <div id="aboutUser">
               <p>
                 Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse voluptate, laudantium, corrupti consectetur repellat eaque rem autem ipsum
