--- js/api/posts/profile/fetch-user-posts.mjs
@@ -1,25 +1,0 @@
-import { API_BASE_URL } from "../../../utilities/base-url.mjs";
-
-export async function fetchUserPosts(userName) {
-  try {
-    const token = localStorage.getItem("jwtToken");
-    if (!token) {
-      throw new Error("No authentication token found");
-    }
-    const response = await fetch(`${API_BASE_URL}/profiles/${userName}/posts`, {
-      method: "GET",
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    if (!response.ok) {
-      throw new Error(`HTTP error. Status: ${response.status}`);
-    }
-
-    const posts = await response.json();
-    return posts;
-  } catch (error) {
-    console.error("Failed to fetch posts", error);
-  }
-}
