--- js/api/posts/profile/profile-posts.mjs
@@ -1,1 +1,0 @@
-
