--- js/api/posts/profile/setup-page.mjs
@@ -1,1 +1,0 @@
-
