--- js/api/posts/profile/user-data.mjs
@@ -1,12 +1,0 @@
-export async function loadUserName() {
-  const userName = localStorage.getItem("userName");
-  if (!userName) {
-    throw new Error("No token received");
-  }
-  return userName;
-}
-
-export function setUserName(userName) {
-  const userNameElement = document.getElementById("userName");
-  userNameElement.innerText = userName;
-}
