+++ js/feed.mjs
@@ -2,6 +2,7 @@ import { fetchPosts } from "./api/posts/fetch-posts.mjs";
 import { renderPosts } from "./api/posts/render-posts.mjs";
 import { addFilterEventListener } from "./api/posts/filter-and-search/filter-posts.mjs";
 import { addSearchEventListener } from "./api/posts/filter-and-search/search-posts.mjs";
+import { loadAndDisplayUserName } from "./utilities/user.mjs";
 import "./api/login_signup/logout.mjs";
 
 // Define viewPost function
@@ -11,12 +12,19 @@ export function viewPost(postId) {
 
 // Fetch and render posts upon DOMContentLoaded
 document.addEventListener("DOMContentLoaded", async () => {
-  const posts = await fetchPosts();
-  renderPosts(posts);
+  try {
+    await loadAndDisplayUserName();
+    const posts = await fetchPosts();
+    document.getElementById("postContainer");
 
-  //Call filter function
-  addFilterEventListener(posts, renderPosts);
+    renderPosts(posts, "postContainer");
 
-  //Call search function
-  addSearchEventListener(posts, renderPosts);
+    //Call filter function
+    addFilterEventListener(posts, renderPosts);
+
+    //Call search function
+    addSearchEventListener(posts, renderPosts);
+  } catch (error) {
+    console.error(error.message);
+  }
 });
